XD-3439: Added status to shell support for tasks
This commit refactors the status mapping code to be external to the
status itself so that it can be independent based on implementation
(task vs process and local vs receptor vs cf).  It also addes the status
check to the building of the task resources when listing the currently
deployed tasks.